Work on the following trackers: 2801858: Typecheck Level 0 Modifiers 2801841: Create abstract syntax for Level 0 modifiers 2801788: Parse level 0 invariants 2801842: Create abstract syntax for level 0 invariants
Also added a test interface to run JML2 like tests, it compares the -expected string comparison with the results of the typechecking, although messages are always not the same but can be helpful to develop and test cases initially before we get something working.

Imported invariant typecheck tests from JML2. Some work as is (those with no errors of course), some you need to actually compare by eye to see that it is working, some does not as a result of non level0 features available in the file.